requirements.txt
flask
python-kasa==0.4.0.dev3